MAINT: Move function for star-shaped cell center to more appropriate module.
No functional changes.
